renamed qiskit_simulator extensions to qasm_simulator_cpp, added basic tests
