+++ qiskit/extensions/qasm_simulator_cpp/__init__.py
@@ -19,5 +19,6 @@
 from .load import load
 from .noise import noise
 from .save import save
-from .uzz import uzz
+from .snapshot import snapshot
+from .rzz import rzz
 from .wait import wait
