+++ qiskit/extensions/qasm_simulator_cpp/load.py
