+++ qiskit/extensions/qasm_simulator_cpp/noise.py
