+++ qiskit/extensions/qasm_simulator_cpp/rzz.py
@@ -28,6 +28,7 @@ from qiskit.qasm import _node as node
 
 
 class RZZGate(Gate):
+    """Two-qubit ZZ-rotation gate."""
 
     def __init__(self, theta, ctl, tgt, circ=None):
         """Create new rzz gate."""
