+++ qiskit/extensions/qasm_simulator_cpp/snapshot.py
