+++ setup.py.in
@@ -43,7 +43,7 @@ packages = ["qiskit",
             "qiskit.dagcircuit",
             "qiskit.extensions",
             "qiskit.extensions.standard",
-            "qiskit.extensions.qiskit_simulator",
+            "qiskit.extensions.qasm_simulator_cpp",
             "qiskit.extensions.quantum_initializer",
             "qiskit.mapper",
             "qiskit.qasm",
